2061 uts homebrew offsetofs cause various pointer-cast warnings Reviewed by: Joshua M. Clulow <josh@sysmgr.org> Reviewed by: Jason King <jason.brian.king@gmail.com> Reviewed by: Darren Reed <avalon@coombs.anu.edu.au> Approved by: Garrett D'Amore <garrett@damore.org>

+++ usr/src/uts/common/fs/udfs/udf_vnops.c
@@ -2759,7 +2759,7 @@ ud_putapage(struct vnode *vp,
 	struct buf *bp = NULL;
 	int32_t lbmask;
 	uint64_t isize;
-	int32_t crc_len;
+	uint16_t crc_len;
 	struct file_entry *fe;
 
 	ud_printf("ud_putapage\n");
